rqspinlock: Allow invoking immediate AA deadlock checks
In-kernel consumes of rqspinlock are often subject to cases where the
same lock held by a task or IRQ context can be requested by an NMI
interrupting their critical section. The other scenario is of reentrancy
within the same context. In both of these cases, reentrancy and taking
the same lock from within the critical section typically causes a
deadlock. Thus far, this has been worked around by recursion counts and
explicit percpu nesting counters that are inc/dec around the lock
acquisition.

When the locks are too many, for instance locks for hash table buckets
in BPF hashmap, such counters are instead restricted to a small number
to avoid memory overhead and hashed, introducing potential collisions
and misdetections.

Enable immediate deadlock detection for all in-kernel users of rqspinlock,
by calling check_timeout immediately and invoking AA deadlock detection
logic. The result of the detection remains the same throughout the
waiting period in the slow path. Since this functionality is relied upon
and handled correctly by in-kernel users, we don't print a warning splat
in such a case.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>