[?] bpf: Convert bpf_timer, bpf_wq to rqspinlock
Convert the bpf_spin_lock used to protect bpf_async struct underpinning
both bpf_timer and bpf_wq.

NOTE:

Now bpf_{wq,timer}_cancel_and_free is fallible, there are several
implications of this. In case __bpf_async_cancel_and_free fails to
acquire the lock, we will simply leak the element instead of freeing it.
This will mean that if we hit a deadlock, timers will keep the prog
refcount, which in turn keeps the map refcount alive and keep firing
till eternity.

Existing code does not handle this, because well, if there was a
deadlock, it would just get hang the kernel.

I am getting a feel it's not the best idea to use rqspinlock here? I'm
inclined to drop this patch, but alternatively, solving the case where
they are leaked would require probably marking a bit in the bpf_async
state asking call fired callbacks to simply skip execution.

This means memory is leaked (for timer, prog, map) but no side effects
occur, which is ofcourse better than a deadlock (well, until this keeps
happening and you run out of memory...).

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>