bpf: Introduce verifier support for rqspinlock
Introduce verifier-side support for rqspinlock kfuncs. The first step is
allowing bpf_res_spin_lock type to be defined in map values and
allocated objects, so BTF-side is updated with a new BPF_RES_SPIN_LOCK
field to recognize and validate.

Any object cannot have both bpf_spin_lock and bpf_res_spin_lock, only
one of them (and at most one of them per-object, like before) must be
present. The bpf_res_spin_lock can also be used to protect objects that
require lock protection for their kfuncs, like BPF rbtree and linked
list.

In terms of verifier plumbing, a new LOCK_CONDITION register type is
introduced to capture the return value semantics of these locks. For
now, all handling is hardcoded to satisfy semantics of the locks, but
the idea can be generalized to other conditional results that cause
different effects to the verifier state based on the result.

Everytime a rqspinlock kfunc is called to acquire a lock, a new lock
state of type REF_TYPE_RES_LOCK_COND or REF_TYPE_RES_LOCK_IRQ_COND (if
the irqsave variant is used) is acquired in the reference state.

This type represents a conditional lock state, which cannot provide
protection for anything just yet, but still has potential side effects
(like possibly disabled preemption, IRQs, possibly taken lock).

When a LOCK_CONDITION is checked to be NULL, in the verifier state
representing the true branch, the reference state's type is patched
to the non-COND type (i.e. REF_TYPE_RES_LOCK or REF_TYPE_RES_LOCK_IRQ).
For the false branch's verifier state, the handling is more involved.

For the non-IRQ type, we just need to release the reference state
created for this LOCK_CONDITION. However, the IRQ-type also needs to
destroy the IRQ state it created on the stack to save any potential
flags. Note that by the time the result is returned, the IRQ state has
already been restored, so a stray bpf_local_irq_restore should not be
able to use this IRQ state to perform a restore. A kfunc_class is
introduced to match the right restore kfunc for IRQ state.

The LOCK_CONDITION checks may also happen out-of-order wrt to the lock
acquisition with irqsave variants, thus simply clearing associated IRQ
state will break the LIFO stack property enforced for IRQ restore.
Hence, extra handling needs to be performed to restore the correct
active_irq_id in case such a case occurs.

Since behavior of preemption, lock, or IRQ-saved states is more
constraining, it is valid for the verifier to verify the program in a
state where they may or may not be held by being more conservative
until the LOCK_CONDITION has been checked.

With all this infrastructure, these kfuncs become usable in programs
while satisfying all safety properties required by the kernel.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>