bpf: Add missing size check for BTF-based ctx access
Robert Morris reported the following program type which passes the
verifier:

SEC("struct_ops/bpf_cubic_init")
void BPF_PROG(bpf_cubic_init, struct sock *sk)
{
	asm volatile("r2 = *(u16*)(r1 + 0)");     // verifier should demand u64
	asm volatile("*(u32 *)(r2 +1504) = 0");   // 1280 in some configs
}

The second line may or may not work, but the first instruction shouldn't
pass, as it's a narrow load into the context structure of the struct
ops callback. The code falls back to btf_ctx_access to ensure
correctness and obtaining the types of pointers. Ensure that the size of
the access is correctly checked to be 8 bytes, otherwise the verifier
thinks the narrow load obtained a trusted BTF pointer and will permit
loads/stores as it sees fit.

Perform the check on size after we've verified that the load is for a
pointer field, as for scalar values narrow loads are fine. [KKD: but
this also means we allow "wider" loads on say u32]. Is this something we
should reject as well?

Fixes: 9e15db66136a ("bpf: Implement accurate raw_tp context access via BTF")
Reported-by: Robert Morris <rtm@mit.edu>
Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>