bpf: Introduce typed BPF arenas
Introduce the idea of typed BPF arenas, which are never shared with user
space and logically bound to a BPF arena map instance.

The idea of introducing typed arenas emanates from the need of
protecting trusted kernel objects that are inter-mixed with arena
objects. Since the verifier performs no static analysis at object
granularity for arena objects, it cannot guarantee the lifetime and
mutation safety for any kernel pointers or objects placed within the
arena. Anything that escapes into an arena memory region must be assumed
to be havoc when loaded back, as there are no guarantees preventing
arbitrary mutation. From the verifier's point of view, an arena is
simply a memory region where it makes no assumptions about its contents.

To allow storage of bpf_res_spin_lock in such memory regions, and
potentially kptrs in the future, we lean on one of the universal ways of
solving any problem in computer science, i.e. indirection.

For every trusted object supposed to be stored within the arena (e.g. a
lock, a kptr, etc.), the program's memory allocation logic will allocate
pages to serve such objects from the typed arena, and then store an
index to the object within the typed arena as a reference instead. For
now, all kfuncs operating on such objects will assume a 32-bit index,
but this is appropriate for now as arenas are limited to 4GB and the
minimium object size we're talking about here is a 4-byte spin lock; at
max a program can create a billion spin lock instances. It is unlikely
programs in the near future will need more spin lock objects than that.

It is true that the number of objects shrinks as the size increases,
e.g. kptrs are half a billion at max per-arena, but a different typed
arena is instantiated for each

Regardless, in terms of future prospects where, say, arenas allow sizes
greater than 4GB, we can revisit this decision by introducing kfuncs
that take a 64-bit pointer and sanitize it instead, removing such
limitations. The key thing is that the "descriptor" size is defined at
the interaction with kfuncs, and a bigger size can be supported with new
kfuncs if need be.

These may not be cases where programs demand >4GB memory, but where
programs transparently share the major chunk of the address space with a
user space application (which allocates most of its data out of the
arena) as a way to transparently share state, and applications where
part of the functionality is implemented as a BPF program running in the
kernel. Therefore, the address space size may need to be more than 4GB
with pages allocated anywhere in the broader region.

From the perspective of the program, it allocates pages within the typed
arena and has full control over their memory management, therefore it
does know the complete pointer to a typed arena object, it just has to
convert that to a 32-bit index for now to store and pass it into kfuncs.

Therefore, supporting a bigger descriptor size for such indirect
pointers in arenas is a matter of accepting wider index as part of
kfuncs. To avoid disruption when this happens, all kfuncs will take a
u64 as part of the prototype from the get go, but not permit values
greater than whatever the max index is (which would fit within 32-bits).
As such, programs can continue to use 32-bits realiably as long as
arenas are sized <= 4G, and can continue to rely on this in the future.

One of the benefits of limiting descriptor size to 32-bits is also
reduction in memory overhead due to the indirection by half: a 64-bit
descriptor size would waste 8 extra bytes per-object in the typed arena,
instead of 4 with 32-bit descriptors.

In the future, more control over the actual size of the typed arenas for
a program can be provided through explicit kfuncs, for now the verifier
will create them on demand and size them to the maximum possible size
necessary through the bpf_typed_arena_alloc kfunc.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>