bpf: Introduce verifier support for arena rqspinlock
Introduce support to handle arena rqspinlock kfuncs. The key change with
PTR_TO_ARENA arguments is that we don't care about the actual state of
this pointer or its value, as it will be runtime checked. As such,
limitations on the constant offsets, presence of spin lock within the
region, etc. can be elided in process_spin_lock.

New reference states pertaining to locked and conditional states are
added and used when acquiring and releasing the lock reference state.

The actual new kfuncs are not exposed directly, instead the existing
rqspinlock kfuncs are made polymorphic depending on the type of the
argument passed into them. For the case of arenas, the same kfuncs will
be called, but replaced with their arena specific variants at runtime.

Aditionally, the arena rqspinlock kfuncs require extra arguments for the
base address of the typed arena of locks, and the maximum count of locks
in them. This is done in fixup_kfunc_call.

Note that for the same instruction index, as a consequence, we cannot
support kfunc calls that take arena and non-arena arguments as we do
instruction patching.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>