[?] bpf: Fix page fault handling in arena rqspinlock
This commit implements an incorret, stop-gap workaround for the
possibility of unmapped pages being referenced by the lock index passed
into the arena kfunc.

The reason is that it is non-trivial to protect concurrent page unmapping
from arena_region_free_pages. Below is an explanation of why RCU is not
sufficient.

The core tenet of deferred reclamation of memory using RCU is that
visibility is removed globally prior to the grace period wait in a
writer. In case of pages, once we clear PTE entries, we also additionally
need to flush the active TLB entries on CPUs running BPF programs that
may access locks resident on these pages.

However, there is no way for us to do this safely. The TLB flush (which
effectively will cause PF on next access to lock word) needs to happen
after existing readers are done using the lock. But in the mean time,
new readers on other CPUs can come in and enter the slow path for the
same lock again.

After the __get_kernel_nofault call is made, the slow path assumes the
address is valid, but the TLB flush can arrive after this point and
cause a crash in the slow path. This can happen in a test where a
background thread constantly tries to map and unmap the typed arena
page, eventually racing the unmap after __get_kernel_nofault but before
the slow path has been exited, causing a PF on access to the lock word.

Somehow, the page unmapping code path needs to be made aware of all
users of a given page hosting lock words, and prevent new users from
coming in. It needs to do all of this without heavy synchronization.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>