[?] bpf: Fix incorrect page_cnt_max calculation
The arena page_cnt_max calculation is borked with map_extra = 0, which
is the case for typed arenas. While this is a stop-gap fix allowing
selftests to run, the complete fix is more involved and requires
touching other parts of alloc_pages and free_pages kfuncs. It will be
split out and sent to the list independently, this diff is enough to run
the series for now.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>