bpf: Elide OoO unlock for arena rqspinlock
In rqspinlock, we explicitly do not support out of order unlocking, as
it can cause misdetection of AA deadlocks, and delay the detection of
ABBA deadlocks until the timeout. However, with arena rqspinlocks, we
have a new set of conditions to consider: since the verifier does not
(and cannot, statically) enforce that the same arena pointer is used
to lock and unlock, it is possible for a program to randomly unlock
*any* lock in a typed arena. While this is mostly harmless as typed
arena locks protect their correspodning arena data structures, it can
introduce a potential stall in the slow path of rqspinlock,
specifically when the unlock happens at this point in the code:

	/*
	 * Note: at this point: (val & _Q_PENDING_MASK) == 0, because of the
	 *       above wait condition, therefore any concurrent setting of
	 *       PENDING will make the uncontended transition fail.
	 */
	if ((val & _Q_TAIL_MASK) == tail) {
----->	>>>>> UNLOCK <<<<<<
		if (atomic_try_cmpxchg_relaxed(&lock->val, &val, _Q_LOCKED_VAL))
			goto release; /* No contention */
	}

	/*
	 * Either somebody is queued behind us or _Q_PENDING_VAL got set
	 * which will then detect the remaining tail and queue behind us
	 * ensuring we'll see a @next.
	 */
	set_locked(lock);

	/*
	 * contended path; wait for next if not observed yet, release.
	 */
	if (!next)
		next = smp_cond_load_relaxed(&node->next, (VAL));

The unlock at that point will cause the new owner of the lock to think
somebody is queued behind them, as the cmpxchg of the lock word failed
after observing that it was the tail waiter. To alleviate this, simply
make unlocks a no-op for arena rqspinlock in case OoO unlocks are
detected. This will mean that OoO unlocking leaves the lock in the
locked state, meaning the next acquisition and any future acquisition
will always cause a timeout.

NOTE: We also have the option of choosing a special "poison" value and
resetting the lock word to that, to allow faster errors in such cases.
TBD.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>