selftests/bpf: Add autogenerated tests for raw_tp NULL args
Add tests to ensure that arguments are correctly masked based on their
specified positions, and whether they get marked correctly as maybe null
or scalar.

We modify the selftest config to enable CONFIG_INFINIBAND to test scalar
marking behavior. To minimize the config, we don't test tracepoints in
the table that are not available with the current config + infiniband.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>