bpf: Implement verifier support for rqspinlock
Introduce verifier-side support for rqspinlock kfuncs. The first step is
allowing bpf_res_spin_lock type to be defined in map values and
allocated objects, so BTF-side is updated with a new BPF_RES_SPIN_LOCK
field to recognize and validate.

Any object cannot have both bpf_spin_lock and bpf_res_spin_lock, only
one of them (and at most one of them per-object, like before) must be
present. The bpf_res_spin_lock can also be used to protect objects that
require lock protection for their kfuncs, like BPF rbtree and linked
list.

The verifier plumbing to simulate success and failure cases when calling
the kfuncs is done by pushing a new verifier state to the verifier state
stack which will simulate failure upon calling the kfunc. The path where
success is indicated creates all lock reference state and IRQ state (if
necessary for irqsave variants). In the case of failure, all state
creation is skipped while verifying the kfunc. When marking the return
value for success case, the value is marked as 0, and for the failure
case as [-MAX_ERRNO, -1]. Then, in the program, whenever user checks the
return value as 'if (ret)' or 'if (ret < 0)' the verifier never
traverses such branches for success cases, and would be aware that the
lock is not held in such cases. To track a failed or successful lock
acquisition at a given insn_idx for lock kfunc, verifier state is
augmented with "spin_lock_failed" member which is reset after processing
the kfunc call.

We introduce a kfunc_class state to avoid mixing lock irqrestore kfuncs
with IRQ state created by bpf_local_irq_save.

With all this infrastructure, these kfuncs become usable in programs
while satisfying all safety properties required by the kernel.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>